batch: Move source lineage into matching package
The line-matching package owns mappings and comparisons while refreshed
source lineage remains at the batch package root.

Lineage translates positions between matched source versions, so leaving it
outside matching separates one coordinate operation from the others.

This commit moves source lineage into line_matching and preserves its
selected import path while focused tests still use the flat location.

The next tests-only commit will migrate matching-focused coverage to the
nested paths before the compatibility modules are removed.
